File moves/organization
1. Move FileUtilsTest and IoUtilsTest to the same sub-project as the classes under test
2. Move classes FileUtils and IoUtils from package: games/strategy/io->org.triplea/io
3. Move WrappedInvocationHandler out of subproject 'java-extras' to 'game-core'
