+++ game-core/src/main/java/games/strategy/engine/data/GameData.java
@@ -10,7 +10,6 @@ import games.strategy.engine.framework.GameDataManager;
 import games.strategy.engine.framework.IGameLoader;
 import games.strategy.engine.framework.message.PlayerListing;
 import games.strategy.engine.history.History;
-import games.strategy.io.IoUtils;
 import games.strategy.triplea.TripleA;
 import java.io.IOException;
 import java.io.ObjectInputStream;
@@ -27,6 +26,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import javax.swing.SwingUtilities;
+import org.triplea.io.IoUtils;
 import org.triplea.thread.LockUtil;
 import org.triplea.util.Tuple;
 import org.triplea.util.Version;
