+++ game-core/src/main/java/games/strategy/engine/framework/startup/mc/ClientModel.java
@@ -29,7 +29,6 @@ import games.strategy.engine.framework.startup.ui.PlayerType;
 import games.strategy.engine.framework.ui.background.WaitWindow;
 import games.strategy.engine.message.RemoteName;
 import games.strategy.engine.player.Player;
-import games.strategy.io.IoUtils;
 import games.strategy.net.ClientMessengerFactory;
 import games.strategy.net.CouldNotLogInException;
 import games.strategy.net.IClientMessenger;
@@ -59,6 +58,7 @@ import javax.swing.SwingUtilities;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.extern.java.Log;
+import org.triplea.io.IoUtils;
 import org.triplea.java.Interruptibles;
 import org.triplea.swing.EventThreadJOptionPane;
 import org.triplea.swing.SwingAction;
