+++ game-core/src/main/java/games/strategy/engine/framework/startup/mc/ServerModel.java
@@ -30,7 +30,6 @@ import games.strategy.engine.framework.startup.login.ClientLoginValidator;
 import games.strategy.engine.framework.startup.ui.PlayerType;
 import games.strategy.engine.framework.startup.ui.ServerOptions;
 import games.strategy.engine.message.RemoteName;
-import games.strategy.io.IoUtils;
 import games.strategy.net.IConnectionChangeListener;
 import games.strategy.net.INode;
 import games.strategy.net.IServerMessenger;
@@ -65,6 +64,7 @@ import org.triplea.domain.data.PlayerName;
 import org.triplea.game.chat.ChatModel;
 import org.triplea.game.server.HeadlessGameServer;
 import org.triplea.game.startup.ServerSetupModel;
+import org.triplea.io.IoUtils;
 import org.triplea.java.Interruptibles;
 import org.triplea.swing.SwingAction;
 import org.triplea.util.Version;
