+++ java-extras/src/test/java/org/triplea/io/IoUtilsTest.java
@@ -1,4 +1,4 @@
-package games.strategy.io;
+package org.triplea.io;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
